refactor(experiments): isolate planner-worker prototype
