+++ loopx/experiments/__init__.py
@@ -1,0 +1,1 @@
+"""Opt-in prototypes that are isolated from LoopX core runtime paths."""
