+++ loopx/experiments/planner_worker/__init__.py
@@ -1,0 +1,1 @@
+"""Experimental planner-worker contract, runtime, and provider adapters."""
