+++ loopx/experiments/planner_worker/traex.py
@@ -10,8 +10,8 @@ import subprocess
 from pathlib import Path
 from typing import Any
 
-from ..planner_worker import AdapterTurn, ValidationResult
-from ..planner_worker_runtime import (
+from .contract import AdapterTurn, ValidationResult
+from .runtime import (
     WorkspaceChange,
     WorkspaceFileType,
     run_planner_worker_once,
