+++ scripts/experiments/traex_planner_worker_probe.py
@@ -6,10 +6,10 @@ import json
 import sys
 from pathlib import Path
 
-REPO_ROOT = Path(__file__).resolve().parents[1]
+REPO_ROOT = Path(__file__).resolve().parents[2]
 sys.path.insert(0, str(REPO_ROOT))
 
-from loopx.extensions.traex_planner_worker import (  # noqa: E402
+from loopx.experiments.planner_worker.traex import (  # noqa: E402
     DEFAULT_TRAEX_PLANNER_MODEL,
     DEFAULT_TRAEX_WORKER_MODEL,
     run_traex_planner_worker_probe,
